Add Speech Recognition and Fix Code
**Summary:**

* Renamed file to *index.py* as it isn't good practice to have spaces in a file name
* Added the `getAudio()` function that uses the `speech_recognition` module. You will need to get the file for the module to upload to the pi pico (which I assume you are using), or use `pip install speech_recognition)
* made some "theoretical" functions comments instead of functions
* made comments regarding next code. 

Hope you like this pr!
